chore: improve Tabs exports
